+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
     </activity>
 
     <activity
-        android:name=".MainActivity"
+        android:name=".main.MainActivity"
         android:theme="@style/AppTheme"/>
 
       <receiver android:name=".DeepLinkAppWidgetProvider">
